\name{interface}
\alias{interface}
\title{Graphical package interface}
\description{
This function displays a main navigation window allowing the access to the different package's options in a graphical environment.
}
\usage{
interface()
}
\details{The funbction display the a navigation window with the main menu. Five buttons proposes:
(1) To construct an ID corespondence file.
(2) To format data files : IrefIndex, Intact, Biogrid.
(3) To build the network.
(4) To model the assembly intermediaries of a protein complex.
(5) To identify proteins of a biological process
}
\examples{
#interface()
}
